+++ Intersect (Core)/GameObjects/Maps/MapResourceSpawn.cs
@@ -5,8 +5,8 @@ namespace Intersect.GameObjects.Maps
     public class ResourceSpawn
     {
         public Guid ResourceId;
-        public int X;
-        public int Y;
-        public int Z;
+        public byte X;
+        public byte Y;
+        public byte Z;
     }
 }
