+++ Intersect.Server/Web/RestApi/Routes/V1/PlayerController.cs
@@ -32,9 +32,9 @@ namespace Intersect.Server.Web.RestApi.Routes.V1
 
         public string Reason { get; set; }
 
-        public int X { get; set; }
+        public byte X { get; set; }
 
-        public int Y { get; set; }
+        public byte Y { get; set; }
 
         public Guid MapId { get; set; }
